requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
groq==0.4.1
requests==2.31.0
pillow==10.0.1
numpy==1.25.2
SpeechRecognition==3.10.0
pydub==0.25.1
gunicorn==21.2.0
Flask-Cors==4.0.0